backend-v2/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
python-dotenv==1.0.0
mysqlclient==2.1.1
pytest==7.3.1